annotate genome

+++ PEI/annotate_cRE/annotate_DHS.py
@@ -105,7 +105,8 @@ def sub_stan(type_bed, col_score, dict_in):
 
 
 def standardize_bed(path_in, path_out, type_bed):
-    os.system(f"rm -rf {path_out}")
+    if os.path.exists(path_out):
+        os.system(f"rm -rf {path_out}")
     os.mkdir(path_out)
     os.system(f"cp {os.path.join(path_in, 'metadata.tsv')} "
               f"{os.path.join(path_out, 'metadata.tsv')}")
@@ -145,7 +146,8 @@ def split_ref_bed(ref_file, dict_in):
 
 
 def merge_split_bed(path_in, path_out):
-    os.system(f"rm -rf {path_out}")
+    if os.path.exists(path_out):
+        os.system(f"rm -rf {path_out}")
     os.mkdir(path_out)
     os.system(f"cp {os.path.join(path_in, 'metadata.tsv')} "
               f"{os.path.join(path_out, 'metadata.tsv')}")
@@ -227,9 +229,18 @@ def annotate_dhs_histone(dict_in):
     return
 
 
-def annotate_dhs(path_dhs_in, path_promoter_in, path_h3k27ac_in, path_dhs_out):
+def match_dhs_file(path_ref, list_histone):
+    meta_dhs = pd.read_csv(os.path.join(path_ref, 'metadata.tsv'), sep='\t')
+
+
+    return
+
+
+def annotate_dhs(path_dhs_in, path_promoter_in, path_h3k27ac_in,
+                 path_h3k4me3_in, path_dhs_out):
     path_out_pro = path_dhs_out + '_pro'
-    os.system(f"rm -rf {path_out_pro}")
+    if os.path.exists(path_out_pro):
+        os.system(f"rm -rf {path_out_pro}")
     os.mkdir(path_out_pro)
 
     list_dhs = generate_file_list(path_dhs_in, path_out_pro)
@@ -238,7 +249,27 @@ def annotate_dhs(path_dhs_in, path_promoter_in, path_h3k27ac_in, path_dhs_out):
     pool.map(func_pro, list_dhs)
     pool.close()
 
+    path_out_tmp = path_dhs_out + '_tmp'
+    if os.path.exists(path_out_tmp):
+        os.system(f"rm -rf {path_out_tmp}")
+    os.mkdir(path_out_tmp)
+    list_h3k4me3 = generate_file_list(path_h3k4me3_in, path_out_tmp)
+
     list_h3k27ac = generate_file_list(path_h3k27ac_in, path_dhs_out)
+    df_h3k4me3 = pd.DataFrame(list_h3k4me3,
+                              columns=['path_out', 'file', 'path_in',
+                                       'organ', 'life_stage', 'term'])
+    df_h3k4me3.columns = ['path_out', 'file', 'path_in_1',
+                          'organ', 'life_stage', 'term']
+    df_h3k27ac = pd.DataFrame(list_h3k27ac,
+                              columns=['path_out', 'file', 'path_in',
+                                       'organ', 'life_stage', 'term'])
+    df_h3k27ac.columns = ['path_out', 'file', 'path_in_2',
+                          'organ', 'life_stage', 'term']
+    df_merge = pd.merge(df_h3k4me3, df_h3k27ac,
+                        on=['organ', 'life_stage', 'term'], how='inner')
+
+
     list_h3k27ac_input = []
     for sub_dict in list_h3k27ac:
         if (sub_dict['organ'] != '.') & (sub_dict['life_stage'] != '.'):
